@@ -16,13 +16,13 @@
     "build": "pnpm -r --parallel build",
     "clean": "pnpm -r exec -- rm -rf node_modules",
     "test": "pnpm -r --parallel test",
-    "fix": "biome check --apply .",
+    "fix": "biome check --write .",
     "version": "node ./scripts/version.mjs",
     "publish": "pnpm build && pnpm changeset publish"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.8.3",
-    "@changesets/cli": "^2.27.1",
+    "@changesets/cli": "^2.27.10",
     "@deck.gl/extensions": "catalog:",
     "@deck.gl/mesh-layers": "catalog:",
     "@deck.gl/test-utils": "catalog:",
@@ -32,16 +32,16 @@
     "@pnpm/types": "^11.1.0",
     "@probe.gl/test-utils": "^3.5.0",
     "@svitejs/changesets-changelog-github-compact": "^1.1.0",
-    "esbuild": "^0.19.5",
-    "esno": "^4.0.0",
+    "esbuild": "^0.25.0",
+    "esno": "^4.8.0",
     "gray-matter": "^4.0.3",
     "tap-spec": "^5.0.0",
     "tape": "^5.5.3",
     "tape-catch": "^1.0.6",
     "tape-run": "^11.0.0",
-    "typescript": "^5.5.4",
+    "typescript": "^5.6.2",
     "unbuild": "^2.0.0",
-    "vite": "^5.3.2"
+    "vite": "^6.2.0"
   },
   "devDependenciesMeta": {
     "@esbuild-plugins/node-globals-polyfill": "Necessary to run `tape` with browser-like `tape-run`.",
