@@ -1,10 +1,10 @@
-import CircularProgress from '@material-ui/core/CircularProgress';
-import Divider from '@material-ui/core/Divider';
-import Grid from '@material-ui/core/Grid';
-import Tab from '@material-ui/core/Tab';
-import Tabs from '@material-ui/core/Tabs';
+import CircularProgress from '@mui/material/CircularProgress';
+import Divider from '@mui/material/Divider';
+import Grid from '@mui/material/Grid2';
+import Tab from '@mui/material/Tab';
+import Tabs from '@mui/material/Tabs';
 import React, { useState } from 'react';
-import shallow from 'zustand/shallow';
+import { useShallow } from 'zustand/shallow';
 
 import { GLOBAL_SLIDER_DIMENSION_FIELDS } from '../../constants';
 import {
@@ -58,7 +58,7 @@ const Controller = () => {
     toggleIsOnSetter,
     removeChannel
   ] = useChannelsStore(
-    store => [
+    useShallow(store => [
       store.channelsVisible,
       store.contrastLimits,
       store.colors,
@@ -68,8 +68,7 @@ const Controller = () => {
       store.setPropertiesForChannel,
       store.toggleIsOn,
       store.removeChannel
-    ],
-    shallow
+    ])
   );
   const loader = useLoader();
 
@@ -86,7 +85,7 @@ const Controller = () => {
     pixelValues,
     isViewerLoading
   ] = useViewerStore(
-    store => [
+    useShallow(store => [
       store.channelOptions,
       store.useLinkedView,
       store.use3d,
@@ -97,8 +96,7 @@ const Controller = () => {
       store.removeIsChannelLoading,
       store.pixelValues,
       store.isViewerLoading
-    ],
-    shallow
+    ])
   );
   const metadata = useMetadata();
   const viewSize = useWindowSize();
@@ -191,7 +189,8 @@ const Controller = () => {
       <Tabs
         value={tab}
         onChange={handleTabChange}
-        aria-label="simple tabs example"
+        aria-label="tabs"
+        variant="fullWidth"
         style={{ height: '24px', minHeight: '24px' }}
       >
         <Tab label="Channels" style={{ fontSize: '.75rem', bottom: 12 }} />
@@ -209,7 +208,12 @@ const Controller = () => {
         {!isViewerLoading && !isRgb ? (
           <Grid container>{channelControllers}</Grid>
         ) : (
-          <Grid container justifyContent="center">
+          <Grid
+            container
+            sx={{
+              justifyContent: 'center'
+            }}
+          >
             {!isRgb && <CircularProgress />}
           </Grid>
         )}
