@@ -1,8 +1,8 @@
-import FormControl from '@material-ui/core/FormControl';
-import InputLabel from '@material-ui/core/InputLabel';
-import Select from '@material-ui/core/Select';
+import FormControl from '@mui/material/FormControl';
+import InputLabel from '@mui/material/InputLabel';
+import Select from '@mui/material/Select';
 import React from 'react';
-import shallow from 'zustand/shallow';
+import { useShallow } from 'zustand/shallow';
 
 import { RENDERING_MODES } from '@hms-dbmi/viv';
 import { useImageSettingsStore, useViewerStore } from '../../../state';
@@ -12,16 +12,18 @@ const renderingOptions = Object.values(RENDERING_MODES);
 function RenderingModeSelect() {
   const renderingMode = useImageSettingsStore(store => store.renderingMode);
   const [isViewerLoading, use3d] = useViewerStore(
-    store => [store.isViewerLoading, store.use3d],
-    shallow
+    useShallow(store => [store.isViewerLoading, store.use3d])
   );
   // Empty option allows for displaying the title of the dropdown fully in the UI.
   const options = !use3d ? [...renderingOptions, ''] : renderingOptions;
   return (
-    <FormControl fullWidth>
-      <InputLabel htmlFor="rendering-mode-select">Rendering Mode</InputLabel>
+    <FormControl fullWidth variant="standard">
+      <InputLabel htmlFor="rendering-mode-select" size="small">
+        Rendering Mode
+      </InputLabel>
       <Select
         native
+        size="small"
         onChange={e =>
           useImageSettingsStore.setState({ renderingMode: e.target.value })
         }
