@@ -1,9 +1,9 @@
-import { createStyles, makeStyles } from '@material-ui/core';
-import Grid from '@material-ui/core/Grid';
-import Slider from '@material-ui/core/Slider';
-import Typography from '@material-ui/core/Typography';
+import Grid from '@mui/material/Grid2';
+import Slider from '@mui/material/Slider';
+import Typography from '@mui/material/Typography';
+import { createStyles, makeStyles } from '@mui/styles';
 import React from 'react';
-import shallow from 'zustand/shallow';
+import { useShallow } from 'zustand/shallow';
 
 import {
   useImageSettingsStore,
@@ -32,8 +32,7 @@ const useStyles = makeStyles(theme =>
 
 const Slicer = () => {
   const [xSlice, ySlice, zSlice] = useImageSettingsStore(
-    store => [store.xSlice, store.ySlice, store.zSlice],
-    shallow
+    useShallow(store => [store.xSlice, store.ySlice, store.zSlice])
   );
   const loader = useLoader();
   const use3d = useViewerStore(store => store.use3d);
@@ -64,11 +63,13 @@ const Slicer = () => {
       <Grid
         container
         direction="row"
-        justifyContent="flex-start"
-        alignItems="center"
         key={label}
+        sx={{
+          justifyContent: 'flex-start',
+          alignItems: 'center'
+        }}
       >
-        <Grid item xs={1} style={{ marginBottom: 8 }}>
+        <Grid item size={1} style={{ marginBottom: 8 }}>
           <Typography
             className={!use3d ? classes.disabled : classes.enabled}
             style={{ marginTop: 4 }}
@@ -76,8 +77,9 @@ const Slicer = () => {
             {label}:
           </Typography>
         </Grid>
-        <Grid item xs={11}>
+        <Grid item size={11}>
           <Slider
+            size="small"
             disabled={!use3d}
             className={!use3d ? classes.disabled : classes.enabled}
             value={val}
